fix(report): 기본 리포트의 시험지 시퀀스가 제대로 적용되지 않는 문제 (#89)
- 시퀀스 로직 수정

+++ src/main/java/bgaebalja/exsherpa/examination/service/ExaminationService.java
@@ -6,7 +6,7 @@ import bgaebalja.exsherpa.examination.domain.SubmitResultRequest;
 import java.util.List;
 
 public interface ExaminationService {
-    Long registerResult(SubmitResultRequest submitResultRequest);
+    int registerResult(SubmitResultRequest submitResultRequest);
 
     List<ExaminationHistory> getSolvedExaminationHistories(String email);
 }
